Fix comment on default sample correction:
The variable nosamplecorr is set to default false (o) in this function. But in the documentation at the top it was written that the functionction defaults to true (1).

The default behavior is correct and it would work if people did not specify the variable nosamplecorr. But if people explicitly look-up which one they want they would have choosen the wrong option.